requirements.txt
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==3.0.3
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
Django==3.2.22
django-cors-headers==3.9.0
django-storages==1.11.1
djangorestframework==3.12.4
exceptiongroup==1.1.3
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3.post1
requests==2.26.0
rfc3986==1.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==1.26.17
zope.interface==6.1